Move DraggableVertex to its own file
