+++ brainframe_qt/ui/dialogs/task_configuration/widgets/in_progress_zone_item.py
@@ -2,15 +2,15 @@ from abc import ABC, abstractmethod
 from typing import List, Optional
 
 from PyQt5.QtCore import Qt
-from PyQt5.QtGui import QColor
 
 from brainframe.api import bf_codecs
 
 from brainframe_qt.ui.resources.config import RenderSettings
-from brainframe_qt.ui.resources.video_items.base import CircleItem, VideoItem
+from brainframe_qt.ui.resources.video_items.base import VideoItem
 from brainframe_qt.ui.resources.video_items.zones import ZoneLineItem, ZoneRegionItem
 
 from ..core.zone import Line, Region, Zone
+from .draggable_vertex import DraggableVertex
 
 
 class InProgressZoneItem(VideoItem, ABC):
@@ -62,7 +62,7 @@ class InProgressZoneItem(VideoItem, ABC):
 
         self.refresh_shape()
 
-        vertex_item = _DraggableVertex(vertex, parent=self)
+        vertex_item = DraggableVertex(vertex, parent=self)
         self._vertex_items.append(vertex_item)
 
     def refresh_shape(self) -> None:
@@ -82,12 +82,13 @@ class InProgressZoneItem(VideoItem, ABC):
             zone is not ready to be drawn yet
         """
 
-    def _init_vertex_items(self) -> List['_DraggableVertex']:
+    def _init_vertex_items(self) -> List['DraggableVertex']:
         assert self.zone.coords is not None
 
-        vertex_items: List[_DraggableVertex] = []
+        vertex_items: List[DraggableVertex] = []
         for coord in self.zone.coords:
-            vertex_items.append(_DraggableVertex(coord, parent=self))
+            vertex_item = DraggableVertex(coord, parent=self)
+            vertex_items.append(vertex_item)
 
         return vertex_items
 
@@ -122,27 +123,3 @@ class InProgressLineItem(InProgressZoneItem):
             line_style=self._line_style,
             parent=self
         )
-
-
-class _DraggableVertex(CircleItem):
-    # TODO: Tie to size of scene
-    DEFAULT_RADIUS = 10
-    DEFAULT_BORDER_THICKNESS = 5
-    DEFAULT_COLOR = QColor(200, 50, 50)
-
-    def __init__(self, position: VideoItem.PointType, *, parent: VideoItem):
-        super().__init__(position, color=self.DEFAULT_COLOR,
-                         radius=self.DEFAULT_RADIUS,
-                         border_thickness=self.DEFAULT_BORDER_THICKNESS,
-                         parent=parent)
-
-        # self.setFlag(self.ItemIsSelectable, True)
-        # self.setFlag(self.ItemIsMovable, True)
-
-    # def mouseMoveEvent(self, event: QGraphicsSceneMouseEvent) -> None:
-    #     event.ignore()
-    #     super().mouseMoveEvent(event)
-    #
-    # def mouseReleaseEvent(self, event: QGraphicsSceneMouseEvent) -> None:
-    #     event.ignore()
-    #     super().mouseMoveEvent(event)
